+++ src/lib/components/icons.js
@@ -47,6 +47,7 @@ import sidebarSimple from "../../assets/icons/phosphor/regular/sidebar-simple.sv
 import browser from "../../assets/icons/phosphor/regular/browser.svg?raw";
 import arrowLeft from "../../assets/icons/phosphor/regular/arrow-left.svg?raw";
 import arrowRight from "../../assets/icons/phosphor/regular/arrow-right.svg?raw";
+import arrowsOutCardinal from "../../assets/icons/phosphor/regular/arrows-out-cardinal.svg?raw";
 import arrowDown from "../../assets/icons/phosphor/regular/arrow-down.svg?raw";
 import arrowUp from "../../assets/icons/phosphor/regular/arrow-up.svg?raw";
 import flag from "../../assets/icons/phosphor/regular/flag.svg?raw";
@@ -166,6 +167,7 @@ export const ICONS = {
   "arrow-down": arrowDown,
   "arrow-up": arrowUp,
   "arrow-right": arrowRight,
+  "arrows-out-cardinal": arrowsOutCardinal,
   flag,
   "dots-six-vertical": dotsSixVertical,
   list,